requirements.txt
# Core web framework
fastapi>=0.104.0

# ASGI server for production
uvicorn[standard]>=0.24.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data validation
pydantic>=2.4.0

# TUI framework for CLI
textual>=0.41.0

# HTTP requests for testing
requests>=2.31.0